PIMP-12 - adds method not allowed and unprocessable entity
